requirements.txt
Django==5.1.4
djangorestframework==3.15.2
pillow==11.0.0
python-dotenv==1.0.1
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
black==24.10.0
psycopg2-binary==2.9.10
pytest==8.3.4



